Minor changes
Made it so that if it is being loaded from a file then debug is automatically on and if it is being loaded from a server then it turns debug off. Also made it so that when moved[] is created it uses a for loop instead of listing it out which isn't better in any sense but it is prettier.